+++ include/memoria/core/packed/sseq/sseq_fn/pkd_f_sseq_select_fn.hpp
@@ -24,7 +24,7 @@ namespace memoria {
 template <typename Seq>
 class BitmapSelectFn {
 
-    static const int32_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
+    static const size_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
 
     static_assert(BitsPerSymbol == 1, "BitmapSelectFn<> can only be used with 1-bit sequences");
 
@@ -32,7 +32,7 @@ class BitmapSelectFn {
 public:
     BitmapSelectFn(const Seq& seq): seq_(seq) {}
 
-    SelectResult operator()(int32_t start, int32_t end, int32_t symbol, int64_t rank)
+    SelectResult operator()(size_t start, size_t end, size_t symbol, int64_t rank)
     {
         if (symbol)
         {
@@ -49,7 +49,7 @@ template <typename Seq>
 class SeqSelectFn {
     typedef typename Seq::Value                                                 Value;
 
-    static const int32_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
+    static const size_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
 
     const Seq& seq_;
 
@@ -59,13 +59,13 @@ class SeqSelectFn {
 public:
     SeqSelectFn(const Seq& seq): seq_(seq) {}
 
-    SelectResult operator()(int32_t start, int32_t end, Value symbol, int32_t rank)
+    SelectResult operator()(size_t start, size_t end, Value symbol, size_t rank)
     {
-        int32_t cnt = 0;
+        size_t cnt = 0;
 
         auto buf = seq_.symbols();
 
-        for (int32_t c = start * BitsPerSymbol; c < end * BitsPerSymbol; c += BitsPerSymbol)
+        for (size_t c = start * BitsPerSymbol; c < end * BitsPerSymbol; c += BitsPerSymbol)
         {
             if (TestBits(buf, c, symbol, BitsPerSymbol))
             {
@@ -91,7 +91,7 @@ public:
 template <typename Seq>
 class Seq8SelectFn {
 
-    static const int32_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
+    static const size_t BitsPerSymbol                                              = Seq::BitsPerSymbol;
 
     static_assert(BitsPerSymbol == 8,
             "Seq8SelectFn<> can only be used with 8-bit sequences");
@@ -101,13 +101,13 @@ class Seq8SelectFn {
 public:
     Seq8SelectFn(const Seq& seq): seq_(seq) {}
 
-    SelectResult operator()(int32_t start, int32_t end, int32_t symbol, int32_t rank)
+    SelectResult operator()(size_t start, size_t end, size_t symbol, size_t rank)
     {
-        int32_t cnt = 0;
+        size_t cnt = 0;
 
         auto buf = seq_.symbols();
 
-        for (int32_t c = start; c < end; c++)
+        for (size_t c = start; c < end; c++)
         {
             if (buf[c] == symbol)
             {
